feat: Transform API to pages and versioning
