+++ src/app/(auth)/login/layout.tsx
@@ -1,0 +1,14 @@
+import "../../../globals.css";
+export default function LoginLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <div className="flex items-center justify-center min-h-screen bg-gray-100">
+      <div className="w-full max-w-md p-4 bg-white rounded shadow">
+        {children}
+      </div>
+    </div>
+  );
+}
