--- src/app/components/index.ts
@@ -1,2 +1,0 @@
-export { SearchInputComponent } from "./SearchInputComponent";
-export * as SideBarMenu from "./sideBarMenu";
