+++ src/app/layout.tsx
@@ -1,7 +1,6 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
-import "./globals.css";
-import SideBarMenu from "./components/sideBarMenu";
+import "../globals.css";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -26,22 +25,9 @@ export default function RootLayout({
   return (
     <html lang="en">
       <head>
-        <link
-          rel="icon"
-          href="/icon.png"
-          type="image/png"
-        />
+        <link rel="icon" href="/icon.png" type="image/png" />
       </head>
-      <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased flex flex-col md:grid md:grid-cols-12 gap-4`}
-      >
-        <div className="md:col-span-2">
-          <SideBarMenu />
-        </div>
-        <div className="md:col-span-10 px-4">
-          {children}
-        </div>
-      </body>
+      <body className={`${geistSans.variable} ${geistMono.variable}`}>{children}</body>
     </html>
   );
 }
