+++ src/components/SearchInputComponent.tsx
