+++ src/components/index.ts
@@ -1,0 +1,2 @@
+export { SearchInputComponent } from "./SearchInputComponent";
+export { SideBarMenu } from "./sideBarMenu";
